# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é uma API de exemplo para o curso de Refatoração em Go.",
        "title": "Go Refatoração API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/categorias": {
            "get": {
                "description": "Retorna todas as categorias cadastradas no banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Listar todas as categorias",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Categoria"
                            }
                        }
                    }
                }
            }
        },
        "/categorias/create": {
            "post": {
                "description": "Cria uma categoria baseada no JSON fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Criar uma nova categoria",
                "parameters": [
                    {
                        "description": "Dados da Categoria",
                        "name": "cat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Categoria"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Categoria"
                        }
                    },
                    "400": {
                        "description": "Erro ao decodificar a categoria",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao criar a categoria",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/categorias/delete": {
            "delete": {
                "description": "Deleta uma categoria baseada no ID fornecido via query string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Deletar uma categoria",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Categoria",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria deletada com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID não fornecido ou inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao deletar a categoria",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/categorias/get": {
            "get": {
                "description": "Retorna uma única categoria baseada no ID fornecido via query string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Buscar categoria por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Categoria",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Categoria"
                        }
                    },
                    "400": {
                        "description": "ID não fornecido ou inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/categorias/update": {
            "put": {
                "description": "Atualiza os dados de uma categoria existente baseada no ID no JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Atualizar uma categoria",
                "parameters": [
                    {
                        "description": "Dados da Categoria (deve conter ID)",
                        "name": "cat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Categoria"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Categoria"
                        }
                    },
                    "400": {
                        "description": "Erro ao decodificar a categoria",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar a categoria",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/itens": {
            "get": {
                "description": "Retorna todos os itens cadastrados no banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "itens"
                ],
                "summary": "Listar todos os itens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Iten"
                            }
                        }
                    }
                }
            }
        },
        "/itens/create": {
            "post": {
                "description": "Cria um item baseado no JSON fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "itens"
                ],
                "summary": "Criar um novo item",
                "parameters": [
                    {
                        "description": "Dados do Item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Iten"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Iten"
                        }
                    },
                    "400": {
                        "description": "Erro ao decodificar o item",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao criar o item",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/itens/delete": {
            "delete": {
                "description": "Deleta um item baseado no ID fornecido via query string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "itens"
                ],
                "summary": "Deletar um item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Item",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item deletado com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID não fornecido ou inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao deletar o item",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/itens/get": {
            "get": {
                "description": "Retorna um único item baseado no ID fornecido via query string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "itens"
                ],
                "summary": "Buscar item por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Item",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Iten"
                        }
                    },
                    "400": {
                        "description": "ID não fornecido ou inválido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Item não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/itens/get-code": {
            "get": {
                "description": "Retorna um único item baseado no código fornecido via query string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "itens"
                ],
                "summary": "Buscar item por código",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código do Item",
                        "name": "codigo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Iten"
                        }
                    },
                    "400": {
                        "description": "Código não fornecido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Item não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/itens/update": {
            "put": {
                "description": "Atualiza os dados de um item existente baseado no ID no JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "itens"
                ],
                "summary": "Atualizar um item",
                "parameters": [
                    {
                        "description": "Dados do Item (deve conter ID)",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Iten"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Iten"
                        }
                    },
                    "400": {
                        "description": "Erro ao decodificar o item",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar o item",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Categoria": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "models.Iten": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "preco": {
                    "type": "number"
                },
                "quantidade": {
                    "type": "integer"
                }
            }
        }
    }
}